refactor(embd): move app to folder
